+++ setup.py
@@ -6,7 +6,7 @@ setup(
     packages=['vent', 'vent.core', 'vent.core.file_drop',
               'vent.core.rq_worker', 'vent.extras.rq_dashboard', 'vent.menus',
               'vent.extras.network_tap', 'vent.extras.network_tap.ncontrol',
-              'vent.extras.rmq_es_connector', 'vent.helpers', 'vent.api'],
+              'vent.extras.rmq_es_connector', 'vent.helpers', 'vent.api', 'vent.legacy'],
     install_requires=['docker>=3.7.0', 'npyscreen>=4.10.5', 'pyyaml>=3.13'],
     scripts=['bin/vent'],
     license='Apache License 2.0',
